Avoid returning PLUM_ERR_INVALID_METADATA on values that a format can't handle
The PLUM_ERR_INVALID_METADATA should only be used to show that a metadata
value is, as the name says, invalid. (For instance, that a value that should
be a uint32_t is five bytes long.) Values that are merely out of range for a
specific format should be ignored or adjusted; metadata is always handled in a
"best effort" basis for each format, and formats that simply cannot handle
some particular metadata should do what they can and ignore the rest. (For
example, a frame duration that exceeds the maximum permitted for a specific
format should simply be adjusted to the maximum valid value.)