Remove references to `NULL` in documentation
C23 introduces `nullptr`, a more suitable null pointer value. It will therefore
become bad practice to use `NULL` in newer versions of C once the new standard
is widely implemented.
Notwithstanding, the library is intentionally usable from any version of C
(starting with C99, or C89 if <stdint.h> and 64-bit integers are available),
and therefore `nullptr` itself should be likewise avoided in documentation.

This commit opts for the middle ground, referencing null pointers as such, in
plain English. Code samples continue to use `NULL`: switching to `nullptr` may
be a consideration when C23 becomes widely available and the library is
migrated to it.